move `ETHEREUM_NETWORK` enum into `config/networks/network.d` definition file
